ner-submission/requirements.txt
spacy==3.0
pandas
tira-rest-api-client